LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-expected.txt

Arabic, مثل هذا النص, is right to left.

This test verifies that the text selection appears visually contiguous when selecting across bidi text boundaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isVisuallyContiguous is true
PASS getSelection().toString() is "مثل هذا النص, is right to left"
PASS successfullyParsed is true

TEST COMPLETE
